[New feature] add maskrcnn (#2908)
* add maskrcnn

* fix mask head

* move maskrcnn to models

* add maskrcnn serialTest

+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/MaskHead.scala
@@ -26,7 +26,7 @@ class MaskHead(
   val inChannels: Int,
   val resolution: Int,
   val scales: Array[Float],
-  val samplingRratio: Float,
+  val samplingRratio: Int,
   val layers: Array[Int],
   val dilation: Int,
   val numClasses: Int,
